Fix cross-platform CI failures before Milestone 2.
Install Linux autotools and Ninja in CI, slim OpenCV via vcpkg features, make uniform RGB images deterministically zero across platforms, fix C ABI last-error lifetime, and resolve compiler warnings.

+++ src/image/image_decoder.cpp
@@ -175,7 +175,8 @@ std::optional<DecodedImage> ImageDecoder::decode(std::span<const std::uint8_t> e
     DecodedImage result;
     result.width = rgb.cols;
     result.height = rgb.rows;
-    result.rgb.resize(static_cast<std::size_t>(rgb.total() * rgb.channels()));
+    result.rgb.resize(static_cast<std::size_t>(rgb.total()) *
+                      static_cast<std::size_t>(rgb.channels()));
     if (!rgb.isContinuous()) {
         return std::nullopt;
     }
